Split data import/export into separate files
